generate colors types too
