+++ src/skins/types/colors.tsx
@@ -1,58 +1,3 @@
-export type SkinVariant = 'prominent';
-
-export type KnownSkinName = 'Movistar' | 'O2' | 'O2-classic' | 'Vivo' | 'Telefonica' | 'Blau';
-export type SkinName = KnownSkinName | string;
-
-export type GetKnownSkin = (variant?: SkinVariant) => KnownSkin;
-
-export type FontWeight = 'light' | 'regular' | 'medium' | 'bold';
-
-type TextPresetName = 'text5' | 'text6' | 'text7' | 'text8' | 'text9' | 'text10' | 'cardTitle';
-export type TextPresetsConfig = {
-    [preset in TextPresetName]: {
-        weight: FontWeight;
-    };
-};
-
-type PartialTextPresetsConfig = {
-    [preset in TextPresetName]?: {
-        weight?: FontWeight;
-    };
-};
-
-export type BorderRadiiConfig = {
-    button: string;
-    input: string;
-    container: string;
-    /**
-     * This will be used for DisplayCards and PosterCards only for the moment. One day we'll unify the borderRadius in all the cards to use container.
-     * @deprecated Use `container` instead
-     */
-    legacyDisplay: string;
-    popup: string;
-    checkbox: string;
-    indicator: string;
-    sheet: string;
-    bar: string;
-    avatar: string;
-};
-
-export type Skin = {
-    name: SkinName;
-    colors: Colors;
-    darkModeColors?: Partial<Colors>;
-    textPresets?: PartialTextPresetsConfig;
-    borderRadii?: BorderRadiiConfig;
-};
-
-export type KnownSkin = {
-    name: KnownSkinName;
-    colors: Colors;
-    darkModeColors?: Partial<Colors>;
-    textPresets?: PartialTextPresetsConfig;
-    borderRadii?: BorderRadiiConfig;
-};
-
 export type Colors = {
     // BACKGROUNDS
     appBarBackground: string;
